+++ Assets/OkapiKit/Scripts/Editor/ActionChangeActionStateEditor.cs
@@ -38,7 +38,7 @@ namespace OkapiKit.Editor
                 if (EditorGUI.EndChangeCheck())
                 {
                     serializedObject.ApplyModifiedProperties();
-                    (target as Action).UpdateExplanation();
+                    action.UpdateExplanation();
                 }
             }
         }