+++ Assets/OkapiKit/Scripts/Editor/ActionChangeTransformEditor.cs
@@ -139,7 +139,7 @@ namespace OkapiKit.Editor
                 if (EditorGUI.EndChangeCheck())
                 {
                     serializedObject.ApplyModifiedProperties();
-                    (target as Action).UpdateExplanation();
+                    action.UpdateExplanation();
                 }
             }
         }