+++ Assets/OkapiKit/Scripts/Editor/Deprecated/ActionChangeValue_DeprecatedEditor.cs
@@ -63,7 +63,7 @@ namespace OkapiKit.Editor
                 if (EditorGUI.EndChangeCheck())
                 {
                     serializedObject.ApplyModifiedProperties();
-                    (target as Action).UpdateExplanation();
+                    (target as OkapiElement).UpdateExplanation();
                 }
             }
         }